moved cypress tests to individual directories. configured snapshots to be stored per test.
