+++ cypress.json
@@ -1,4 +1,6 @@
 {
     "baseUrl": "http://localhost:3000",
-    "projectId" : "d944bj"
+    "projectId" : "d944bj",
+    "useRelativeSnapshots": true,
+    "snapshotFileName": "snapshot.data"
 }
