+++ cypress/integration/home_page/home_page.spec.js
@@ -1,0 +1,34 @@
+describe('The Home Page', () => {
+
+    beforeEach(() => {
+        cy.viewport(1024, 768);
+        cy.visit('/');
+    });
+
+    it('successfully loads', () => {
+        cy.checkTitle('Digital Public Library of America');
+        cy.contains('images, texts, videos, and sounds from across the United States');
+        cy.contains('Online Exhibitions');
+        cy.contains('Primary Source Sets');
+        cy.contains('How can I use DPLA?');
+        cy.contains('Education');
+        cy.contains('Family Research');
+        cy.contains('Lifelong Learning');
+        cy.contains('Scholarly Research');
+        cy.contains('View all user guides');
+        cy.contains('Get the latest DPLA news in your inbox');
+
+        cy.getDataCy('home-hero').within(() => {
+            cy.getDataCy('dpla-logo').snapshot();
+            cy.get('form').snapshot();
+        });
+
+        cy.getDataCy('stay-informed').snapshot();
+        cy.getDataCy('guides').snapshot();
+
+        cy.getDataCy('Exhibitions').should('exist');
+        cy.getDataCy("Sets").should('exist');
+        cy.getDataCy('news').should('exist');
+        cy.getDataCy('link-home').should('not.exist');
+    });
+  });
