--- cypress/integration/topics.spec.js
@@ -1,32 +1,0 @@
-const topics = require('../fixtures/topics.json')
-describe('Topics', async () => {
-
-    beforeEach(() => {
-        cy.viewport(1024, 768);
-    });
-
-    it(`home hasn't changed`, () => {
-        cy
-            .visit('/browse-by-topic')
-            .getDataCy('topics-home')
-            .snapshot();
-    });
-
-    for (const topic of topics) {
-        it(`${topic.slug}'s home hasn't changed`, () => {
-            cy
-                .visit(`/browse-by-topic/${topic.slug}`)
-                .getDataCy('topic-home')
-                .snapshot();
-        });
-
-        for (const subtopic of topic.subtopics) {
-            it(`${topic.slug} ${subtopic} works correctly`, () => {
-                cy.visit(`/browse-by-topic/${topic.slug}/${subtopic}`);
-                cy.getDataCy('subtopic-sidebar').snapshot();
-                cy.getDataCy('subtopic-item-list').find('li').its('length').should('be.gt', 5);
-            });
-        }
-    }
-});
-
