+++ cypress/plugins/index.js
@@ -1,3 +1,5 @@
+const fs = require('fs')
+
 /// <reference types="cypress" />
 // ***********************************************************
 // This example plugins/index.js can be used to load plugins
@@ -17,6 +19,16 @@
  */
 // eslint-disable-next-line no-unused-vars
 module.exports = (on, config) => {
-  // `on` is used to hook into various events Cypress emits
-  // `config` is the resolved Cypress config
+  on('task', {
+    // Needs to be done through a task, as we can't access `fs` in test case
+    // (this task runs in Node which allows use of external libraries).
+    readFileMaybe(filename) {
+      // console.log('Loading ', filename);
+      if (fs.existsSync(filename)) {
+        return fs.readFileSync(filename, 'utf8');
+      }
+      // console.log('Could not find ', filename);
+      return '';
+    }
+  });
 }
